Use WaypointSearch in tl_detector
Create a new module for the common waypoint search method,
implemented in waypoint_updater, and use this also in tl_detector.
Currently this is however implemented in two equal submodules due to
problems with import from other ROS packages.

The waypoint search functionality in tl_detector is also slightly
modified to find the waypoint closest ahead of ego vehicle, and the
waypoint closest behind each stop-line.

This change also corrects a wrap-around bug that otherwise would make
the program fail in finding next TL at the end of a lap, e.g. when
next TL is in the beginning of next lap.
