+++ ros/src/tl_detector/tl_detector.py
@@ -1,19 +1,18 @@
 #!/usr/bin/env python
 import rospy
 from std_msgs.msg import Int32
-from geometry_msgs.msg import PoseStamped, Pose
+from geometry_msgs.msg import PoseStamped
 from styx_msgs.msg import TrafficLightArray, TrafficLight
 from styx_msgs.msg import Lane
 from sensor_msgs.msg import Image
 from cv_bridge import CvBridge
 from light_classification.tl_classifier import TLClassifier
+from waypoint_search import WaypointSearch
 import tf
-import cv2
 import yaml
-import numpy as np
-from scipy.spatial import KDTree
 
 STATE_COUNT_THRESHOLD = 3
+
 
 class TLDetector(object):
     def __init__(self):
@@ -24,8 +23,8 @@ class TLDetector(object):
         self.camera_image = None
         self.lights = []
 
-        sub1 = rospy.Subscriber('/current_pose', PoseStamped, self.pose_cb)
-        sub2 = rospy.Subscriber('/base_waypoints', Lane, self.waypoints_cb)
+        rospy.Subscriber('/current_pose', PoseStamped, self.pose_cb)
+        rospy.Subscriber('/base_waypoints', Lane, self.waypoints_cb)
 
         '''
         /vehicle/traffic_lights provides you with the location of the traffic light in 3D map space and
@@ -34,8 +33,8 @@ class TLDetector(object):
         simulator. When testing on the vehicle, the color state will not be available. You'll need to
         rely on the position of the light and the camera image to predict it.
         '''
-        sub3 = rospy.Subscriber('/vehicle/traffic_lights', TrafficLightArray, self.traffic_cb)
-        sub6 = rospy.Subscriber('/image_color', Image, self.image_cb)
+        rospy.Subscriber('/vehicle/traffic_lights', TrafficLightArray, self.traffic_cb)
+        rospy.Subscriber('/image_color', Image, self.image_cb)
 
         config_string = rospy.get_param("/traffic_light_config")
         self.config = yaml.load(config_string)
@@ -51,9 +50,6 @@ class TLDetector(object):
         self.last_wp = -1
         self.state_count = 0
 
-        self.waypoints_2d = None
-        self.waypoints_tree = None
-        self.n_waypoints = 0
         rospy.spin()
 
     def pose_cb(self, msg):
@@ -61,10 +57,7 @@ class TLDetector(object):
 
     def waypoints_cb(self, waypoints):
         self.waypoints = waypoints
-        # Preprocess to make it faster to search for closest waypoint using the k-d tree algorithm.
-        self.waypoints_2d = [[waypoint.pose.pose.position.x, waypoint.pose.pose.position.y]
-                             for waypoint in self.waypoints.waypoints]
-        self.waypoints_tree = KDTree(self.waypoints_2d)
+        self.wp_search = WaypointSearch(self.waypoints.waypoints)
 
     def traffic_cb(self, msg):
         self.lights = msg.lights
@@ -77,7 +70,7 @@ class TLDetector(object):
             msg (Image): image from car-mounted camera
 
         """
-        if self.waypoints_tree is None:
+        if None in (self.pose, self.wp_search):
             return
 
         self.has_image = True
@@ -102,19 +95,6 @@ class TLDetector(object):
             self.upcoming_red_light_pub.publish(Int32(self.last_wp))
         self.state_count += 1
 
-    def get_closest_waypoint(self, x, y):
-        """Identifies the closest path waypoint to the given position
-            https://en.wikipedia.org/wiki/Closest_pair_of_points_problem
-        Args:
-            pose (Pose): position to match a waypoint to
-
-        Returns:
-            int: index of the closest waypoint in self.waypoints
-
-        """
-        closest_idx = self.waypoints_tree.query([x, y], 1)[1]
-        return closest_idx
-
     def get_light_state(self, light):
         """Determines the current color of the traffic light
 
@@ -131,7 +111,7 @@ class TLDetector(object):
 
         cv_image = self.bridge.imgmsg_to_cv2(self.camera_image, "bgr8")
 
-        #Get classification
+        # Get classification
         return self.light_classifier.get_classification(cv_image)
 
     def process_traffic_lights(self):
@@ -148,16 +128,17 @@ class TLDetector(object):
 
         # List of positions that correspond to the line to stop in front of for a given intersection
         stop_line_positions = self.config['stop_line_positions']
-        if(self.pose):
-            car_wp_idx = self.get_closest_waypoint(self.pose.pose.position.x, self.pose.pose.position.y)
+        car_wp_idx = self.wp_search.get_closest_waypoint_idx_ahead(self.pose.pose.position.x,
+                                                                   self.pose.pose.position.y)
 
         #TODO find the closest visible traffic light (if one exists)
         diff = len(self.waypoints.waypoints)
         for i, light in enumerate(self.lights):
-            line = stop_line_positions[i]
-            temp_wp_idx = self.get_closest_waypoint(line[0], line[1])
-            d = temp_wp_idx - car_wp_idx
-            if d >= 0 and d < diff:
+            stop_line_x, stop_line_y = stop_line_positions[i]
+            temp_wp_idx = self.wp_search.get_closest_waypoint_idx_behind(stop_line_x,
+                                                                         stop_line_y)
+            d = (temp_wp_idx - car_wp_idx) % len(self.waypoints.waypoints)
+            if d < diff:
                 diff = d
                 closest_light = light
                 line_wp_idx = temp_wp_idx
@@ -167,6 +148,7 @@ class TLDetector(object):
             return line_wp_idx, state
 
         return -1, TrafficLight.UNKNOWN
+
 
 if __name__ == '__main__':
     try:
