+++ ros/src/waypoint_updater/waypoint_updater.py
@@ -5,12 +5,11 @@ from copy import deepcopy
 from std_msgs.msg import Bool
 from geometry_msgs.msg import PoseStamped, TwistStamped
 from styx_msgs.msg import Lane, TrafficLightArray, TrafficLight, Waypoint
-from scipy.spatial import KDTree
 import numpy as np
 import math
 import yaml
 from speed_calculator import SpeedCalculator
-
+from waypoint_search import WaypointSearch
 '''
 This node will publish waypoints from the car's current position to some `x` distance ahead.
 
@@ -286,42 +285,6 @@ class WaypointCalculator(object):
                 # velocity can be limited to the max allowed value here.
                 rospy.logerr('Calculated velocity %s exceeds max allowed value %s.', velocity, self.max_velocity)
                 set_waypoint_velocity(self.waypoints, i, self.max_velocity)
-
-
-class WaypointSearch(object):
-    """Organize waypoints to make it fast to search for the closest to a position"""
-    def __init__(self, waypoints):
-        # Preprocess waypoints using the k-d tree algorithm
-        self.waypoints_2d = [[wp.pose.pose.position.x, wp.pose.pose.position.y] for wp in waypoints]
-        self.waypoints_tree = KDTree(self.waypoints_2d)
-
-    def get_closest_waypoint_idx_ahead(self, x, y):
-        closest_idx = self.get_closest_waypoint_idx(x, y)
-
-        if not self.__is_closest_idx_ahead(closest_idx, x, y):
-            closest_idx = (closest_idx + 1) % len(self.waypoints_2d)
-        return closest_idx
-
-    def get_closest_waypoint_idx_behind(self, x, y):
-        closest_idx = self.get_closest_waypoint_idx(x, y)
-        if self.__is_closest_idx_ahead(closest_idx, x, y):
-            closest_idx = (closest_idx - 1) % len(self.waypoints_2d)
-        return closest_idx
-
-    def get_closest_waypoint_idx(self, x, y):
-        return self.waypoints_tree.query([x, y], 1)[1]
-
-    def __is_closest_idx_ahead(self, closest_idx, x, y):
-        """ Check if closest_idx is ahead or behind position"""
-        closest_coord = self.waypoints_2d[closest_idx]
-        prev_coord = self.waypoints_2d[closest_idx - 1]
-
-        # Equation for hyperplane through closest_coord
-        cl_vect = np.array(closest_coord)
-        prev_vect = np.array(prev_coord)
-        pos_vect = np.array([x, y])
-        val = np.dot(cl_vect - prev_vect, pos_vect - cl_vect)
-        return val < 0
 
 
 # Helper functions
